requirements.txt
numpy==2.0.2
scikit-learn==1.6.1
torch==2.9.0
torchvision==0.24.0
tensorflow==2.19.0
pandas==2.3.3
pillow==11.3.0
torchinfo==1.8.0
matplotlib==3.10.0